requirements.txt
assemblyai
python-dotenv
streamlit       

